requirements.txt
arabic-reshaper
asgiref==3.2.10
asn1crypto==1.3.0
bcrypt==3.1.7
cffi
cryptography
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap3==12.0.3
django-easy-pdf==0.1.1
django-mail-templated==2.6.5
django-utils-six==2.0
django-widget-tweaks==1.4.8
fabric==2.5.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.9
invoke==1.4.1
#mysqlclient==1.4.6
paramiko==2.7.1
Pillow
pyasn1==0.4.8
pycparser==2.20
PyNaCl
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple
pytz==2019.3
reportlab
six==1.14.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.5
